+++ reqbaz/src/main/java/de/rwth/dbis/acis/bazaar/service/BazaarService.java
@@ -39,6 +39,8 @@ import de.rwth.dbis.acis.bazaar.service.notification.NotificationDispatcher;
 import de.rwth.dbis.acis.bazaar.service.notification.NotificationDispatcherImp;
 import de.rwth.dbis.acis.bazaar.service.resources.*;
 import de.rwth.dbis.acis.bazaar.service.security.AuthorizationManager;
+import de.rwth.dbis.acis.bazaar.service.twitter.TweetDispatcher;
+import de.rwth.dbis.acis.bazaar.service.twitter.WeeklyNewProjectsTweetTask;
 import i5.las2peer.api.Context;
 import i5.las2peer.api.ManualDeployment;
 import i5.las2peer.api.ServiceException;
@@ -50,6 +52,7 @@ import i5.las2peer.logging.L2pLogger;
 import i5.las2peer.restMapper.RESTService;
 import i5.las2peer.restMapper.annotations.ServicePath;
 import io.swagger.annotations.*;
+import lombok.Getter;
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.apache.http.client.utils.URIBuilder;
 import org.jooq.SQLDialect;
@@ -66,7 +69,6 @@ import javax.ws.rs.core.Response;
 import javax.xml.bind.DatatypeConverter;
 import java.net.HttpURLConnection;
 import java.net.URISyntaxException;
-import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.*;
@@ -88,6 +90,7 @@ public class BazaarService extends RESTService {
     private final ValidatorFactory validatorFactory;
     private final List<BazaarFunctionRegistrar> functionRegistrar;
     private final NotificationDispatcher notificationDispatcher;
+    private final TweetDispatcher tweetDispatcher;
     private final DataSource dataSource;
 
     //CONFIG PROPERTIES
@@ -103,6 +106,10 @@ public class BazaarService extends RESTService {
     protected String smtpServer;
     protected String emailFromAddress;
     protected String emailSummaryTimePeriodInMinutes;
+    protected String twitterApiKey;
+    protected String twitterApiKeySecret;
+    protected String twitterClientId;
+    protected String twitterClientSecret;
 
     public BazaarService() throws Exception {
         setFieldValues();
@@ -139,6 +146,8 @@ public class BazaarService extends RESTService {
             notificationDispatcher.setActivityDispatcher(new ActivityDispatcher(this, activityTrackerService, activityOrigin, baseURL, frontendBaseURL));
         }
 
+        Timer timer = new Timer();
+
         if (!smtpServer.isEmpty()) {
             Properties props = System.getProperties();
             props.put("mail.smtp.host", smtpServer);
@@ -147,7 +156,7 @@ public class BazaarService extends RESTService {
             if (!emailSummaryTimePeriodInMinutes.isEmpty()) {
                 try {
                     // This task is scheduled to run every 60 seconds * emailSummaryTimePeriodInMinutes
-                    Timer timer = new Timer();
+
                     timer.scheduleAtFixedRate((NotificationDispatcherImp) notificationDispatcher, 0, 60000 * Integer.valueOf(emailSummaryTimePeriodInMinutes));
                 } catch (Exception ex) {
                     logger.warning(ex.getMessage());
@@ -155,6 +164,10 @@ public class BazaarService extends RESTService {
             }
 
         }
+
+        tweetDispatcher = new TweetDispatcher(this, twitterApiKey, twitterApiKeySecret, twitterClientId, twitterClientSecret);
+
+        new WeeklyNewProjectsTweetTask(this).schedule(timer);
 
         notificationDispatcher.setBazaarService(this);
     }
@@ -181,6 +194,7 @@ public class BazaarService extends RESTService {
         //getResourceConfig().register(PersonalisationDataResource.class);
         getResourceConfig().register(FeedbackResource.class);
         getResourceConfig().register(WebhookResource.class);
+        getResourceConfig().register(AdminResource.class);
     }
 
     public String getBaseURL() {
@@ -216,6 +230,10 @@ public class BazaarService extends RESTService {
 
     public NotificationDispatcher getNotificationDispatcher() {
         return notificationDispatcher;
+    }
+
+    public TweetDispatcher getTweetDispatcher() {
+        return tweetDispatcher;
     }
 
     private void registerUserAtFirstLogin() throws Exception {
@@ -281,6 +299,10 @@ public class BazaarService extends RESTService {
 
     public ObjectMapper getMapper() {
         return mapper;
+    }
+
+    public String getFrontendBaseURL() {
+        return frontendBaseURL;
     }
 
     public Response.ResponseBuilder paginationLinks(Response.ResponseBuilder responseBuilder, PaginationResult paginationResult,